Use sinf and cosf in surgescript_transform_rotate2d()

+++ util/transform.c
@@ -84,7 +84,7 @@ void surgescript_transform_translate2d(surgescript_transform_t* t, float x, floa
 void surgescript_transform_rotate2d(surgescript_transform_t* t, float degrees)
 {
     float a = degrees * DEG2RAD;
-    float s = sin(a), c = cos(a);
+    float s = sinf(a), c = cosf(a);
     t->rotation.sz = t->rotation.sz * c + t->rotation.cz * s;
     t->rotation.cz = t->rotation.cz * c - t->rotation.sz * s;
 }
